requirements.txt
# Core ML framework
torch

# Hugging Face ecosystem
transformers>=4.44.0
peft>=0.12.0
accelerate>=0.33.0
datasets
huggingface-hub

# Serialization / data
safetensors
pandas
requests
tqdm

# Image processing
Pillow
pillow-heif
